// src/pages/mypage/profile/edit/edit.types.ts
export interface MyPageForm {
  positions: string[];
  projectTypes: string[];
  skills: string[];
  languages: string[];
  personalities: string[];
  job: string;
  areas: string[];
  imageUrl: File;
  name: string;
  introduce: string;
  blog: string | null;
}

export interface EditProfileProps {
  userId: string;
}